More robust queueing technique, log to stdout.
Now uses Queue everywhere instead of deque for the output
queue, the container itself isn't as needed as I thought previously.

Changed default logger stream handler to write to stdout.

Instead of waiting for the threads to join, use a more stable
method of pulling and waiting for the threads to complete on a timeout
and then skipping if they aren't fast enough. Leads to much more
proper and predictable running, and the inexplicable deadlocks
that happen when a child thread is joined from the parent.